pyproject.toml
[project]
name = "jiuge"
version = "2.0.0-alpha"
description = "Jiuge - 个人音乐 AI 助手"
authors = [{name = "DaimaRuge"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["music", "ai", "agent", "spotify", "netease", "assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # 音乐平台 API
    "NeteaseCloudMusicApi>=0.1.0",
    "spotipy>=2.23.0",
    "apple-music-python>=0.1.0",
    
    # 音频播放
    "ffpyplayer>=4.3.0",
    
    # 音乐元数据
    "musicbrainzngs>=0.7.1",
    "mutagen>=1.47.0",
    
    # 推荐算法
    "implicit>=0.7.2",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    
    # 异步支持
    "aiohttp>=3.8.0",
    "httpx>=0.24.0",
    
    # Web API
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
    
    # 配置管理
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    
    # 数据存储
    "aiosqlite>=0.19.0",
    
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # 缓存
    "redis>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.4.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.1.0",
]

[project.scripts]
jiuge = "jiuge.cli:app"

[project.urls]
Homepage = "https://github.com/DaimaRuge/Jiuge"
Documentation = "https://github.com/DaimaRuge/Jiuge#readme"
Repository = "https://github.com/DaimaRuge/Jiuge"
Issues = "https://github.com/DaimaRuge/Jiuge/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
